requirements.txt
absl-py==1.0.0
amqp==5.1.1
async-timeout==4.0.2
autopep8==1.6.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
cachetools==5.0.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cycler==0.11.0
Cython==0.29.28
decorator==4.4.2
Deprecated==1.2.13
docopt==0.6.2
easydict==1.9
filelock==3.6.0
flake8==4.0.1
Flask==2.1.1
flask-sock==0.5.2
fonttools==4.31.2
future==0.18.2
gdown==4.4.0
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.3
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
kombu==5.2.4
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
moviepy==1.0.3
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
proglog==0.1.10
prompt-toolkit==3.0.29
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==4.1.1
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy==1.8.0
seaborn==0.11.2
simple-websocket==0.5.2
six==1.16.0
soupsieve==2.3.1
tb-nightly==2.9.0a20220331
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.1
Turbo-Flask==0.8.0
typing-extensions==4.1.1
urllib3==1.26.9
vext==0.7.6
vext.gi==0.7.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.1
wsproto==1.1.0
yacs==0.1.8
yapf==0.32.0
zipp==3.7.0
psycopg2-binary==2.9.3